refactor(dashboard)
It's a hero dashboard, so it should go inside hero/
