+++ src/modules/routing/routes.ts
@@ -1,5 +1,5 @@
 import { Routes } from '@angular/router';
-import { DashboardComponent } from 'src/modules/app/components/dashboard/dashboard.component';
+import { DashboardComponent } from 'src/modules/app/hero/components/dashboard/dashboard.component';
 import { HeroesComponent } from 'src/modules/app/hero/components/heroes/heroes.component';
 import { HeroDetailComponent } from 'src/modules/app/hero/components/detail/hero-detail.component';
 import { HelloWorldComponent } from '../app/components/hello-world/hello-world.component';
